Move ThumbnailX to Core lib
