+++ src/IIIFPresentation/Core.Tests/Usings.cs
@@ -1,0 +1,2 @@
+global using FluentAssertions;
+global using Xunit;
